ForesToolboxRS: Remote Sensing Tools for Forest Monitoring

Remote sensing tools for forest monitoring includes a series of novel techniques to monitor forest through optical images and fusion of optical and microwave data. All tools are based on scientific publications.

See also

Author

Maintainer: Yonatan Tarazona Coronel geoyons@gmail.com (ORCID)

Authors:

Jakub Nowosad nowosad.jakub@gmail.com (ORCID)

Other contributors:

Alaitz Zabala alaitz.zabala@uab.cat (ORCID) [contributor]

Xavier Pons xavier.pons@uab.cat (ORCID) [contributor]

Antoni Broquetas broquetas@tsc.upc.edu (ORCID) [contributor]

Hamdi A. Zurqani hzurqan@clemson.edu (ORCID) [contributor]