Normalized Difference Fraction Index (NDFI)

The NDFI it is sensitive to the state of the canopy cover, and has been successfully applied to monitor forest degradation and deforestation in Peru and Brazil. This index comes from the Green Vegetation (GV), non-photosynthetic vegetation (NPV), Soil (S) and the reminder is the shade component.

ndfiSMA(img, procesLevel = "SR", verbose = FALSE)

Arguments

img

It could be RasterStack or RasterBrick

procesLevel

Processing level. It is possible to obtain the NDFI from images in surface reflectance (SR) from TM, ETM+ and OLI, or Top of Atmosphere (TOA) values only for Landsat 8 OLI. The default is SR. In addition, for any processing level, the image values must be rescaled between 0 and 10000

verbose

This paramater is Logical. It Prints progress messages during execution

Note

The fractions is obtained from the Spectral Mixture Analysis physical model.

References

Souza Jr., C.M., Roberts, D.A., Cochrane, M.A., 2005. Combining spectral and spatialinformation to map canopy damage from selective logging and forest fires. Remote Sens. Environ. 98 (2-3), 329-343.

Examples

library(ForesToolboxRS)
library(raster)

# Load an example dataset
data(img_l8)

# Unmix the image
ndfi <- ndfiSMA(img = img_l8, procesLevel = "SR")
plot(ndfi)