Change detection using the PVts-\(\beta\) approach (raster version)

This algorithm allows to detect disturbances in the forests using Landsat data. It can also be run on data from other sensors such as MODIS.

pvtsRaster(x, startm, endm, threshold = 5, img, vf = FALSE, verbose = FALSE)

Arguments

x
Matrix, RasterStack or Rasterbrick without NA's

startm
The start of the monitoring time

endm
The end of the monitoring time

threshold
The default threshold is 5 for photosynthetic vegetation or for Normalized Difference Fraction Index (NDFI), while for indices such as NDVI and EVI the threshold is 3. Please see Tarazona et al. (2018) for more details.

img
The image of the position immediately before the monitoring start, i.e. the "start-1" position (in case "x" is a matrix). The matrix x and the image img must have the same dimension (i.e. rows and cols).

vf
If the monitoring is with Photosynthetic Vegetation series, then switch to TRUE

verbose
This parameter is Logical. It Prints progress messages during execution

References

Tarazona, Y., Mantas, V.M., Pereira, A.J.S.C. (2018). Improving tropical deforestation detection through using photosynthetic vegetation time series (PVts-\(\beta\)). Ecological Indicators, 94, 367 379.

Note

In order to optimize the detection, it is advisable to make a smoothing before detecting changes. So the matrix or the stack must go through a smoothing first. The smoothing will allow to eliminate outliers that were not eliminated during the masking. See smootH for more details. In addition, in case the input is a matrix, the first dimension must be rows*columns of the image, and the second dimension the number of images.

Examples

if (FALSE) {
library(ForesToolboxRS)
library(raster)
data(serie_pv)

# Detect changes in 2008 (position 19) using a raster(RasterStack)
cd <- pvtsRaster(x = serie_pv, startm = 19, endm = 19, threshold = 5)
}